requirements.txt
fastapi
uvicorn[standard]
litellm
python-dotenv
colorlog
rich
httpx
